+++ src/components/general-widgets/MultiRoomType/utils/getPluralString.js
@@ -1,0 +1,8 @@
+/**
+ * @param {Number} amount
+ * @param {String} singular
+ * @param {String} plural
+ * @return {String}
+ */
+export const getPluralString = (amount, singular, plural) =>
+  `${amount} ${amount === 1 ? singular : plural}`;